+++ plugins/idea/src/main/kotlin/org/pastalab/fray/idea/ui/FrayDebugPanel.kt
@@ -8,17 +8,18 @@ import com.intellij.openapi.project.Project
 import java.awt.BorderLayout
 import javax.swing.JPanel
 import javax.swing.JSplitPane
-import org.pastalab.fray.idea.debugger.FrayScheduleObserver
 import org.pastalab.fray.idea.getPsiFile
-import org.pastalab.fray.idea.mcp.SchedulerMcpExplorer
+import org.pastalab.fray.idea.getPsiFileFromClass
 import org.pastalab.fray.idea.objects.ThreadExecutionContext
+import org.pastalab.fray.mcp.ClassSourceProvider
+import org.pastalab.fray.mcp.SchedulerDelegate
+import org.pastalab.fray.mcp.SchedulerMcpExplorer
+import org.pastalab.fray.rmi.ScheduleObserver
+import org.pastalab.fray.rmi.ThreadInfo
 import org.pastalab.fray.rmi.ThreadState
 
-class FrayDebugPanel(
-    val project: Project,
-    val scheduleObserver: FrayScheduleObserver,
-    replayMode: Boolean
-) : JPanel() {
+class FrayDebugPanel(val project: Project, replayMode: Boolean) :
+    JPanel(), ScheduleObserver<ThreadExecutionContext> {
   // UI Components
   private val controlPanel: SchedulerControlPanel
   private val threadTimelinePanel: ThreadTimelinePanel
@@ -26,8 +27,8 @@ class FrayDebugPanel(
 
   // Data management
   private val threadInfoUpdaters: MutableMap<Editor, ThreadInfoUpdater> = mutableMapOf()
-  var selected: ThreadExecutionContext? = null
-  private var callback: ((ThreadExecutionContext) -> Unit)? = null
+  var selected: ThreadInfo? = null
+  private var callback: ((ThreadInfo) -> Unit)? = null
   val highlightManager = MultiThreadHighlightManager()
 
   init {
@@ -43,7 +44,6 @@ class FrayDebugPanel(
 
     // Create the thread timeline panel
     threadTimelinePanel = ThreadTimelinePanel()
-    scheduleObserver.observers.add(threadTimelinePanel)
 
     // Create the thread resource panel
     threadResourcePanel = ThreadResourcePanel()
@@ -57,7 +57,19 @@ class FrayDebugPanel(
     add(mainSplitPane, BorderLayout.CENTER)
   }
 
-  private val mcpServer = SchedulerMcpExplorer(project, controlPanel)
+  private val mcpServer =
+      SchedulerMcpExplorer(
+          object : ClassSourceProvider {
+            override fun getClassSource(className: String): String? {
+              return getPsiFileFromClass(className, project)?.text
+            }
+          },
+          object : SchedulerDelegate {
+            override fun scheduled(thread: ThreadInfo) {
+              scheduleButtonPressed(thread)
+            }
+          },
+          replayMode)
 
   private fun createRightPanel(): JPanel {
     // Create a panel to hold timeline and resource panels
@@ -73,7 +85,7 @@ class FrayDebugPanel(
     return rightPanel
   }
 
-  fun scheduleButtonPressed(newSelected: ThreadExecutionContext?) {
+  fun scheduleButtonPressed(newSelected: ThreadInfo?) {
     ApplicationManager.getApplication().invokeAndWait { highlightManager.clearAll() }
     threadInfoUpdaters.forEach { it.value.threadNameMapping.clear() }
     threadInfoUpdaters.clear()
@@ -93,19 +105,28 @@ class FrayDebugPanel(
     threadInfoUpdaters.clear()
     controlPanel.clear()
     threadResourcePanel.clear()
-    scheduleObserver.observers.remove(threadTimelinePanel)
   }
 
+  /**
+   * Schedules the next operation for the given threads.
+   *
+   * @param threads The list of `ThreadExecutionContext` representing the threads to be scheduled.
+   * @param scheduled The `ThreadExecutionContext` of the thread that is currently scheduled, `null`
+   *   in explore mode.
+   * @param onThreadSelected A callback function to be invoked when a thread is selected.
+   */
   fun schedule(
       threads: List<ThreadExecutionContext>,
-      onThreadSelected: (ThreadExecutionContext) -> Unit
+      scheduled: ThreadExecutionContext?,
+      onThreadSelected: (ThreadInfo) -> Unit
   ) {
-    mcpServer.newSchedulingRequestReceived(threads)
+    mcpServer.newSchedulingRequestReceived(threads.map { it.threadInfo }, scheduled?.threadInfo)
     processThreadsForHighlighting(threads)
 
     // We need to update the control panel after highlighting because
     // highlighting changes the opened editors, and we need to switch back.
-    controlPanel.updateThreads(threads, selected)
+    controlPanel.updateThreads(
+        threads, threads.firstOrNull { it.threadInfo.threadIndex == selected?.threadIndex })
 
     // Update thread resource information
     threadResourcePanel.updateThreadResources(threads)
@@ -143,6 +164,23 @@ class FrayDebugPanel(
       }
     }
   }
+
+  override fun onExecutionStart() {
+    mcpServer.onExecutionStart()
+  }
+
+  override fun onNewSchedule(
+      allThreads: List<ThreadExecutionContext>,
+      scheduled: ThreadExecutionContext
+  ) {
+    threadTimelinePanel.onNewSchedule(allThreads, scheduled)
+  }
+
+  override fun onExecutionDone(bugFound: Throwable?) {
+    mcpServer.onExecutionDone(bugFound)
+  }
+
+  override fun saveToReportFolder(path: String) {}
 
   companion object {
     const val CONTENT_ID = "fray-scheduler"